+++ CHANGELOG.md
@@ -6,8 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Experimental
+### Changed
+- Take Save out of "more" menu
+
+## [1.112] - 2022-11-25
 ### Added
-- Native player for video-files
+- Group administrators can now block users in their managed groups. A blocked
+  user cannot post to the group, but can read and comment if the group is not
+  private.
+- Video attachments now have a player, when browser supports them
+- YouTube shorts now supported in media viewer
+
+### Fixed
+- No refresh needed to view private users and groups after subscription approved
+- No refresh needed to interact with new subscription requests
+- Hidden comment class name updated to avoid interference with Firefox builtin extension style
 
 ## [1.111.2] - 2022-09-23
 ### Fixed